Move to servlets.pages package. Move Dashboard user management.
