+++ flightmate/src/main/java/com/flightmate/servlets/pages/LoginServlet.java
@@ -1,4 +1,4 @@
-package com.flightmate.servlets;
+package com.flightmate.servlets.pages;
 
 import java.io.IOException;
 import java.util.Iterator;
