+++ flightmate/src/main/java/com/flightmate/servlets/pages/LogoutServlet.java
@@ -1,4 +1,4 @@
-package com.flightmate.servlets;
+package com.flightmate.servlets.pages;
 
 import java.io.IOException;
